Use vocab cutoff to select terms for fitting the model

The same text, speaker, and filter should be used in this model as in fit_speaker to select the terms for the latter function.

stylest_terms(x, speaker, vocab_cutoff, filter = NULL)

Arguments

x
Corpus as text vector. May be a corpus_frame object

speaker
Vector of speaker labels. Should be the same length as x

vocab_cutoff
Quantile cutoff for the vocabulary in (0, 100]

filter
if not NULL, a corpus filter

Value

list of terms

Examples

data(novels_excerpts)
stylest_terms(novels_excerpts$text, novels_excerpts$author, vocab_cutoff = 50)

#>    [1] ","               "the"             "of"              "."              
#>    [5] "and"             "in"              "to"              "a"              
#>    [9] "for"             ";"               "as"              "that"           
#>   [13] "i"               "with"            "it"              "had"            
#>   [17] "be"              "on"              "was"             "my"             
#>   [21] "or"              "but"             "his"             "have"           
#>   [25] "when"            "been"            "not"             "who"            
#>   [29] "at"              "by"              "he"              "is"             
#>   [33] "one"             "so"              "there"           "from"           
#>   [37] "very"            "all"             "-"               "her"            
#>   [41] "their"           "me"              "which"           "an"             
#>   [45] "little"          "into"            "this"            "even"           
#>   [49] "she"             "them"            "years"           "mr"             
#>   [53] "any"             "are"             "day"             "has"            
#>   [57] "if"              "may"             "some"            "they"           
#>   [61] "do"              "long"            "many"            "only"           
#>   [65] "up"              "will"            "great"           "where"          
#>   [69] "more"            "you"             "were"            "life"           
#>   [73] "men"             "about"           "own"             "what"           
#>   [77] "ago"             "can"             "just"            "lady"           
#>   [81] "most"            "say"             "we"              "before"         
#>   [85] "well"            "am"              "days"            "him"            
#>   [89] "man"             "must"            "never"           "no"             
#>   [93] "should"          "than"            "young"           "few"            
#>   [97] "good"            "know"            "like"            "made"           
#>  [101] "much"            "perhaps"         "three"           "too"            
#>  [105] "\""              "!"               "shall"           "?"              
#>  [109] "mother"          ":"               "among"           "came"           
#>  [113] "certain"         "county"          "dear"            "died"           
#>  [117] "family"          "father"          "himself"         "interest"       
#>  [121] "lived"           "present"         "spoke"           "these"          
#>  [125] "upon"            "wife"            "after"           "age"            
#>  [129] "down"            "end"             "every"           "existence"      
#>  [133] "first"           "house"           "indeed"          "known"          
#>  [137] "land"            "last"            "nothing"         "other"          
#>  [141] "park"            "people"          "position"        "short"          
#>  [145] "single"          "surrounding"     "tell"            "think"          
#>  [149] "time"            "two"             "under"           "whom"           
#>  [153] "without"         "world"           "your"            "miss"           
#>  [157] "our"             "girl"            "how"             "now"            
#>  [161] "then"            "daughters"       "death"           "handsome"       
#>  [165] "kind"            "large"           "mistress"        "old"            
#>  [169] "principal"       "small"           "subject"         "thousand"       
#>  [173] "town"            "travelled"       "'"               "("              
#>  [177] ")"               "acquaintance"    "acquainted"      "affairs"        
#>  [181] "affectionate"    "afternoon"       "almost"          "appeared"       
#>  [185] "barely"          "beginning"       "between"         "bred"           
#>  [189] "cease"           "centuries"       "child"           "connection"     
#>  [193] "considered"      "course"          "dark"            "deep"           
#>  [197] "desire"          "did"             "early"           "enough"         
#>  [201] "enter"           "entering"        "eyes"            "families"       
#>  [205] "far"             "fond"            "fortune"         "four"           
#>  [209] "friend"          "friends"         "given"           "grandfather"    
#>  [213] "half"            "heard"           "here"            "history"        
#>  [217] "home"            "hope"            "hour"            "hours"          
#>  [221] "impatiently"     "inhabitants"     "invitation"      "journey"        
#>  [225] "lad"             "lawyer"          "legal"           "line"           
#>  [229] "longer"          "look"            "make"            "might"          
#>  [233] "mrs"             "myself"          "nearly"          "once"           
#>  [237] "over"            "point"           "poor"            "pretty"         
#>  [241] "property"        "proud"           "quite"           "read"           
#>  [245] "receive"         "remembrance"     "respect"         "respectable"    
#>  [249] "saying"          "sea-coast"       "seen"            "seventeen"      
#>  [253] "sir"             "sister"          "sisters"         "situation"      
#>  [257] "sixteen"         "such"            "sure"            "taken"          
#>  [261] "things"          "though"          "thought"         "towards"        
#>  [265] "turned"          "twenty"          "twice"           "unless"         
#>  [269] "village"         "whose"           "woman"           "would"          
#>  [273] "year"            "same"            "saw"             "above"          
#>  [277] "bennet"          "estate"          "god"             "letters"        
#>  [281] "vernon"          "want"            "—"               "back"           
#>  [285] "broad"           "christmas"       "consumption"     "cursed"         
#>  [289] "dashwood"        "dog"             "drop"            "eltham"         
#>  [293] "emma"            "england"         "events"          "exceptional"    
#>  [297] "farm"            "forgive"         "glendwr"         "governess"      
#>  [301] "huntingdon"      "ink"             "knew"            "make-believe"   
#>  [305] "maria"           "married"         "miseries"        "national"       
#>  [309] "norland"         "occurred"        "oh"              "owain"          
#>  [313] "post"            "pounds"          "prejudices"      "shire"          
#>  [317] "spinning-wheels" "to-day"          "truth"           "ward"           
#>  [321] "welsh"           "wilkins"         "1492"            "1799"           
#>  [325] "1836"            "_angina"         "_them_"          "_you_"          
#>  [329] "able"            "abroad"          "abrupt"          "acknowledged"   
#>  [333] "across"          "actors"          "add"             "admiration"     
#>  [337] "admitted"        "advanced"        "advantage"       "affection"      
#>  [341] "affectionately"  "against"         "alien"           "alien-looking"  
#>  [345] "allowed"         "almonds"         "alpine"          "already"        
#>  [349] "alteration"      "always"          "ambition"        "amusement"      
#>  [353] "angel"           "annie"           "announced"       "another"        
#>  [357] "answer"          "anyhow"          "approaches"      "arising"        
#>  [361] "arrange"         "ask"             "assertion"       "attacks"        
#>  [365] "attending"       "attorney"        "aware"           "away"           
#>  [369] "baby"            "bad"             "bag"             "barked"         
#>  [373] "baronet's"       "baronetage"      "became"          "begin"          
#>  [377] "being"           "believe"         "bells"           "benefited"      
#>  [381] "bent"            "bequeath"        "bertram"         "best"           
#>  [385] "better"          "blankness"       "bless"           "blessings"      
#>  [389] "blindly"         "blood"           "bloom"           "blushing"       
#>  [393] "book"            "books"           "born"            "bosom"          
#>  [397] "both"            "bound"           "branch"          "brawny"         
#>  [401] "breakfasts"      "bringing"        "british"         "broken"         
#>  [405] "brother"         "builder"         "burden"          "burge"          
#>  [409] "burnt"           "burthen"         "busily"          "business"       
#>  [413] "busy"            "buy"             "called"          "calling"        
#>  [417] "captivate"       "career"          "caresses"        "carpenter"      
#>  [421] "carrying"        "case"            "catholics"       "cattle"         
#>  [425] "caucasus"        "caught"          "centre"          "century"        
#>  [429] "certainly"       "chance"          "changed"         "character"      
#>  [433] "charge"          "cheerful"        "chieftain"       "children"       
#>  [437] "church"          "churchhill"      "cities"          "claim"          
#>  [441] "clear"           "clergyman"       "clerk"           "clever"         
#>  [445] "clock"           "close"           "clothed"         "coaches"        
#>  [449] "come"            "comer"           "comes"           "comfortable"    
#>  [453] "comforts"        "coming"          "companion"       "conceal"        
#>  [457] "conduct"         "confectioner"    "confectioner's"  "confectionery"  
#>  [461] "confidential"    "connected"       "consequence"     "consequences"   
#>  [465] "considerable"    "consideration"   "consolation"     "consolidated"   
#>  [469] "constant"        "constitution"    "contained"       "contemplating"  
#>  [473] "contempt"        "continental"     "contrast"        "convenient"     
#>  [477] "conveyancing"    "corn"            "could"           "country"        
#>  [481] "country-folk"    "countrymen"      "couple"          "creations"      
#>  [485] "cried"           "cumberland"      "dare"            "daresay"        
#>  [489] "daughter"        "dawn"            "dead"            "deafen"         
#>  [493] "dears"           "debts"           "dec"             "degree"         
#>  [497] "delightful"      "delivering"      "delusive"        "deserve"        
#>  [501] "desired"         "desolation"      "different"       "dines"          
#>  [505] "discern"         "discretion"      "disinherited"    "disposition"    
#>  [509] "dispositions"    "distinction"     "distress"        "distressed"     
#>  [513] "districts"       "domes"           "domestic"        "doors"          
#>  [517] "dressed-up"      "driven"          "dumplings"       "eager"          
#>  [521] "earliest"        "earthly"         "egyptian"        "eighteenth"     
#>  [525] "elevation"       "elliot"          "else"            "endless"        
#>  [529] "endured"         "engage"          "engineer"        "enlightenment"  
#>  [533] "envy"            "equal"           "equally"         "equitable"      
#>  [537] "established"     "ever"            "excellent"       "excitement"     
#>  [541] "exclaimed"       "expectation"     "experience"      "expressed"      
#>  [545] "extraordinary"   "faculties"       "failed"          "faint"          
#>  [549] "fairly"          "fallen"          "famous"          "far-away"       
#>  [553] "far-reaching"    "farmhouses"      "fatalities"      "favourable"     
#>  [557] "fed"             "feeling"         "feelings"        "fell"           
#>  [561] "fiercely"        "fifteen"         "figure"          "fills"          
#>  [565] "firm"            "five"            "fix"             "fixed"          
#>  [569] "flash"           "folks"           "foresee"         "forests"        
#>  [573] "fortitude"       "forty"           "forward"         "found"          
#>  [577] "frances"         "free"            "full"            "fully"          
#>  [581] "general"         "generality"      "generations"     "gentry"         
#>  [585] "get"             "gives"           "glass"           "glazing"        
#>  [589] "glendower"       "go"              "got"             "grandmother"    
#>  [593] "grass"           "greatness"       "green"           "groan"          
#>  [597] "guidance"        "guys"            "hall"            "hamley"         
#>  [601] "hands"           "happened"        "happy"           "hayslope"       
#>  [605] "health"          "hearing"         "hearts"          "heaven"         
#>  [609] "heavy"           "henry"           "hercules"        "hero"           
#>  [613] "hills"           "honest"          "hornby"          "hospitable"     
#>  [617] "household"       "housekeeper"     "however"         "human"          
#>  [621] "hummed"          "husband"         "idle"            "ill"            
#>  [625] "illiterate"      "immediately"     "improvements"    "incidents"      
#>  [629] "income"          "independence"    "indistinct"      "indulgent"      
#>  [633] "inexperienced"   "influential"     "inheritor"       "inroads"        
#>  [637] "inside"          "insular"         "intended"        "interested"     
#>  [641] "intermediate"    "intimacy"        "introduced"      "invited"        
#>  [645] "isles"           "james"           "jars"            "jerky"          
#>  [649] "jews"            "jonathan"        "joy"             "june"           
#>  [653] "kellynch"        "kept"            "knowledge"       "ladies"         
#>  [657] "lanes"           "langford"        "late"            "lately"         
#>  [661] "lead"            "leaf"            "learnt"          "lease"          
#>  [665] "least"           "leaving"         "left"            "leigh"          
#>  [669] "less"            "let"             "levant"          "likes"          
#>  [673] "limited"         "lips"            "live"            "lodgings"       
#>  [677] "looked"          "looking"         "lord"            "loss"           
#>  [681] "love"            "lozenges"        "luck"            "lucy"           
#>  [685] "ludlow"          "macaroons"       "magical"         "makes"          
#>  [689] "making"          "manner"          "mansfield"       "many-curved"    
#>  [693] "marriage"        "marrying"        "masts"           "match"          
#>  [697] "measurer"        "meet"            "men's"           "mental"         
#>  [701] "meringues"       "mid"             "mighty"          "miles"          
#>  [705] "mind"            "minds"           "mingled"         "minutes"        
#>  [709] "mirror"          "modify"          "money"           "month"          
#>  [713] "months"          "morning"         "motion"          "mountain"       
#>  [717] "mounts"          "multitude"       "mysterious"      "nakedness"      
#>  [721] "name"            "natives"         "naturally"       "need"           
#>  [725] "needed"          "needle"          "neighbourhood"   "nephew"         
#>  [729] "netherfield"     "new"             "north"           "northampton"    
#>  [733] "notes"           "nought"          "nurse-maid"      "oak"            
#>  [737] "objection"       "obliged"         "obtained"        "occupation"     
#>  [741] "off"             "offends"         "offer"           "often"          
#>  [745] "olive"           "one-and-twenty"  "opened"          "opinion"        
#>  [749] "order"           "ordinary"        "orphan"          "otherwise"      
#>  [753] "out"             "outline"         "outweigh"        "owen"           
#>  [757] "owner"           "oxford"          "pains"           "pale"           
#>  [761] "pallid"          "paradise"        "parents"         "parted"         
#>  [765] "particularly"    "passing"         "past"            "pastry-cook's"  
#>  [769] "patents"         "patriotism"      "pay"             "peasant"        
#>  [773] "pectoris_"       "pen"             "peppermint"      "period"         
#>  [777] "person"          "physical"        "physician"       "pillars"        
#>  [781] "pine"            "pink"            "pitch"           "pity"           
#>  [785] "place"           "places"          "plain"           "plains"         
#>  [789] "pleasure"        "plum-buns"       "poem"            "poignancy"      
#>  [793] "polished"        "pork"            "possession"      "powerless"      
#>  [797] "powers"          "precepts"        "predict"         "presence"       
#>  [801] "pressing"        "pretend"         "prince"          "principality"   
#>  [805] "principally"     "prize"           "prizes"          "produced"       
#>  [809] "professional"    "profiting"       "prolong"         "prosper"        
#>  [813] "protestants"     "protracted"      "proudly"         "provide"        
#>  [817] "provision"       "provisions"      "purchase"        "pure"           
#>  [821] "puritans"        "race"            "races"           "rain-freshened" 
#>  [825] "raised"          "raising"         "rank"            "rarely"         
#>  [829] "rather"          "rattling"        "reach"           "reader"         
#>  [833] "readers"         "received"        "redoubted"       "refuge"         
#>  [837] "refuse"          "relating"        "remnant"         "remnants"       
#>  [841] "rented"          "replied"         "repugnance"      "residence"      
#>  [845] "restless"        "retirement"      "returned"        "reveal"         
#>  [849] "rich"            "ridges"          "rightful"        "ringing"        
#>  [853] "rise"            "rising"          "river-sides"     "rochdale"       
#>  [857] "room"            "roomy"           "round"           "roused"         
#>  [861] "sad"             "said"            "salt"            "sate"           
#>  [865] "satiety"         "satirized"       "satisfied"       "scarcely"       
#>  [869] "scattered"       "scholar"         "school"          "science"        
#>  [873] "scotland"        "screaming"       "scruple"         "sea"            
#>  [877] "secure"          "sedge-like"      "seemed"          "seems"          
#>  [881] "sensations"      "sentence"        "separation"      "serve"          
#>  [885] "service"         "settled"         "seven"           "shadows"        
#>  [889] "sharp"           "shepherd's"      "shop"            "show"           
#>  [893] "showing"         "side"            "sidereal"        "sign"           
#>  [897] "silk"            "sister's"        "six"             "skill"          
#>  [901] "slightest"       "slow"            "snowy"           "soaring"        
#>  [905] "socially"        "society"         "sold"            "somersetshire"  
#>  [909] "something"       "somewhere"       "son"             "soon"           
#>  [913] "sorcerer"        "sorrow"          "spelling"        "spending"       
#>  [917] "spires"          "spoken"          "spots"           "spring-time"    
#>  [921] "standing"        "stars"           "start"           "state"          
#>  [925] "station"         "stay"            "stayed"          "steady"         
#>  [929] "steal"           "still"           "stirred"         "stock"          
#>  [933] "stomach"         "stooped"         "story"           "strange"        
#>  [937] "strangely"       "strengthened"    "strict"          "strongly"       
#>  [941] "studied"         "sufficiently"    "sugar-candy"     "sugared"        
#>  [945] "summits"         "sun"             "sunset"          "supplied"       
#>  [949] "supply"          "suppose"         "sups"            "susan"          
#>  [953] "sussex"          "taking"          "tale"            "taylor"         
#>  [957] "tedium"          "tells"           "ten"             "tender"         
#>  [961] "thanks"          "thee"            "thereby"         "therefore"      
#>  [965] "thing"           "thirty"          "thomas"          "those"          
#>  [969] "thou'rt"         "thread-lace"     "three-cornered"  "times"          
#>  [973] "told"            "took"            "tortured"        "toy"            
#>  [977] "traditions"      "transacted"      "true"            "try"            
#>  [981] "twelfth-cake"    "twenty-one"      "unceasing"       "uncle"          
#>  [985] "undersized"      "undertake"       "undertaken"      "undertakes"     
#>  [989] "unit"            "unite"           "universally"     "unstable"       
#>  [993] "unwelcome"       "unwilling"       "upland"          "upright"        
#>  [997] "urgent"          "us"              "utter"           "valleys"        
#> [1001] "vex"             "views"           "visions"         "wainscoting"    
#> [1005] "wales"           "walk"            "walter"          "warm"           
#> [1009] "watching"        "way"             "wearisome"       "week"           
#> [1013] "weeks"           "western"         "westmoreland"    "whether"        
#> [1017] "while"           "whisper"         "whistle"         "whizz"          
#> [1021] "widow"           "window-frames"   "wing"            "winter"         
#> [1025] "wisdom"          "wise"            "witches"         "within"         
#> [1029] "wizards"         "women"           "woodhouse"       "woodhouse's"    
#> [1033] "words"           "workmen"         "workshop"        "ye"             
#> [1037] "yeast"           "yeoman"          "yet"             "youngest"       
#> [1041] "youth"           "“"               "”"